+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix SpireField to work with generic types
 * Fix SpireField to not use duplicate objects
 * Fix NPE in isModLoaded
+* Force defining paramtypes on overloaded methods
 
 #### v2.9.1 ####
 * Patch to always enable Custom mode